Improve exocc command line interface
Adds two output modes:

  1. -o/--outdir OUTDIR [-s/--stem STEM]
  2. -oc/--cfile CFILE -oh/--hfile HFILE

In the first mode, the user specifies at least an OUTDIR for
the .c and .h files. If they are compiling only one Exo source
file, then the STEM is taken from that file. Otherwise, STEM
must be provided.

In the second mode, the user specifies exact paths for the
.c and .h files as CFILE and HFILE, respectively. This is
to support build systems (like Bazel and Meson) whose rules
for relating input file names to output file names for code
generators are more stringent.

Additionally, adds --debug DEBUG option where a DEBUG mode
may be specified. For now, only `schedule` is understood
as a debug mode. When specified, the top-level funcs are
printed in their final, scheduled form. In a follow-up PR,
this will be extended to the transitive closure of all
procs that will be compiled.

Fixes #264
Fixes #270