Make FormattedExpr objects non-reusable
The object is designed as a wrapper to pass an argument. It shouldn't
be reusable. If we keep cursors inside it, we risk the cursors getting
forwarded. Instead we store the LoopIR objects. The object is likely not
reusable after a scheduling operation, but if the user keep a reference
to it, we are keeping the LoopIR objects within for no reason.