Convert several scheduling directives to use cursor editing APIs (#323, #333)
## Scheduling directives

* cut_loop
* mult_loops
* insert_pass
* delete_config
* merge_writes
* unroll
* inline
* remove_loop

Note that these may have sub-optimal forwarding policies.

## Pattern matching

The pattern_match function can now limit its search to the scope beneath
a given cursor.

## Internal cursors

Remove knowledge of Procedures from internal cursors.

The internal cursor editing functions used to return a new Procedure
(api-level) and a forwarding function. Now they return a new LoopIR.proc
(internal) and a forwarding function.

This change was necessary to support delete_config, which requires also
changing the _mod_config property on the Procedure, in addition to the
provenance.

Now that the internal cursors do not reference API Procedures (well,
almost... that's the end goal), the weak-reference optimization became
less profitable, so I removed it for the sake of simplicity. It used to
be that leaking a single cursor could keep an entire provenance chain
alive. Now it can only keep one tree alive.