Forward gaps and internal cursor cleanup (#406)
Implement gap forwarding and add a simple test.

Also, clean-up a few more LoopIR references from the internal cursors:

* Move internal cursor's `children()` method to `pattern_match`
* Rename `_get_loopir()` to `resolve_all()`

Now, only `_delete` has a reference to LoopIR (for inserting passes).
The LoopIR invariants should change. Philosophically, the _type_ of a
for loop or if-statement body isn't `list`, but
`non-empty-statement-list`. Practically, it will be easier to interpret
empty lists the same way as `pass`.