Fix normalization and range analysis bugs
When merging a binop range, range analysis was implicitly assuming that
those ranges are completely independent i.e. the expression contains
only one read of each symbol that occurs in the expression. This is now
an explicit check.

Moreover, _DoNormalize had a related bug where we
were calling range analysis on the Seq hi expression before nornalizing
it.

Finally, there was an unrelated bug, which this commit fixes,
were the symbols ranges environment in _DorNormalize wasn't being
properly setup in if/else statement.