Adds parallel loops and scheduling op to parallelize loops (#526)
This is motivated by the effort to support Halide's `compute_at` and
`store_at` commands in Exo. The observation is that for blur, these
operations are only useful when used **with** parallel threads.

As part of this chance, we renamed `Seq` to `For` so that we can
differentiate between sequential and parallel for loops. This will also
require modifying the BLAS repository accordingly to reflect that.

Some future considerations: 
- Do we ever want a scheduling operation to revert back to sequential
loops?
- Different types of parallelism that we want to represent – the current
one just generates `#pragma omp parallel for`.