Added support for uint16 and improved `replace_all` to handle bodies with length > 1 (#536)
This commit adds a precision for `uint16_t` to Exo, which was necessary
to compare against the Halide blur benchmark. I also added the
corresponding AVX2 instructions that were necessary for the blur effort.
One important system change is that `replace_all` is now capable of
handling instructions whose proc semantics are longer than a single
statement.