Unify index inequalities with different ops
* Since index inequalities are operating on integers, we can simply
rewrite all of them from the form: `lhs op rh` to `0 < expr`. Then, we
can perform unification on the two resulting rhs expressions. This will
allow for more unification oppurtunities.
* We have previously implemented an subcase of this, but required the
two operations to be the same. This commit generalizes this idea to
unify two inequalities of arbitrary ops against each other. Equalities
can still only be unified if the other operation is also an equality.