Updated Halide schedule to use circular buffer optimization (#629)
In order to make the circular buffer optimization work, I had to modify
`compute_at_with_prologue`. This is due to our limited range analysis
being insufficient to reason about for loops with an `i == 0` clause in
the body, so I commented out the code which fuses the prologue into the
main loop body.