WIP: Partial window of window support (second-order window) (#762)
This is not a complete implementation. Full support would require
changing scheduling operators to recognize windows as tensors (e.g. as a
source for stage_mem) and/or also have to avoid creating invalid
T.Window types or AttributeError-ing due to treating a T.Window as if it
were a T.Tensor.

The key helper function is `LoopIR.create_window_type` which takes care
of "inlining" the indexing effects of two windows together.

Fixes https://github.com/exo-lang/exo/issues/594

---------

Co-authored-by: Yuka Ikarashi <yukatkh@gmail.com>